+++ ChangeLog.md
@@ -6,6 +6,8 @@
 
 - Relaxed the type of the parameters for `GeminiURI.__init__`.
   ([#24](https://github.com/davep/wasat/pull/24))
+- Added `GeminiURI.with_default_scheme`
+  ([#25](https://github.com/davep/wasat/pull/25))
 
 ## v0.5.0